+++ fabric.ts
@@ -14,8 +14,10 @@ export type {
   TCanvasSizeOptions,
   TSVGExportOptions,
 } from './src/canvas/StaticCanvas';
+export type { StaticCanvasOptions } from './src/canvas/StaticCanvasOptions';
 export { StaticCanvas } from './src/canvas/StaticCanvas';
 export { Canvas } from './src/canvas/Canvas';
+export type { CanvasOptions } from './src/canvas/CanvasOptions';
 export { CanvasDOMManager } from './src/canvas/DOMManagers/CanvasDOMManager';
 export { StaticCanvasDOMManager } from './src/canvas/DOMManagers/StaticCanvasDOMManager';
 
@@ -41,7 +43,6 @@ export { PatternBrush } from './src/brushes/PatternBrush';
 
 export { FabricObject as Object } from './src/shapes/Object/FabricObject';
 export type {
-  TProps,
   TFabricObjectProps,
   FabricObjectProps,
   SerializedObjectProps,